requirements.txt
certifi==2017.11.5
chardet==3.0.4
clearbit==0.1.7
Django==1.11.6
django-cors-headers==2.1.0
django-jet==1.0.7
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
idna==2.6
pyhunter==0.3
PyJWT==1.5.3
pytz==2017.3
requests==2.18.4
urllib3==1.22
whitenoise==3.3.1
